refactor: consolidate modules and organize scripts
