+++ scripts/cache_daily_data.py
