+++ scripts/tickers_loader.py
