+++ scripts/update_from_bulk_last_day.py
