+++ strategies/system1_strategy.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 import pandas as pd
 
 from .base_strategy import StrategyBase
-from system.core import (
+from core.system1 import (
     prepare_data_vectorized_system1,
     generate_roc200_ranking_system1,
     get_total_days_system1,
