--- system/__init__.py
@@ -1,8 +1,0 @@
-"""
-System-wide shared utilities and core logic.
-
-This package exposes common functions used across multiple systems.
-"""
-
-# Re-export commonly used names for convenience if needed later.
-
