--- tmp_import_check.py
@@ -1,3 +1,0 @@
-import importlib
-m = importlib.import_module('run_all_systems_today')
-print('import_ok')
