--- utils/__init__.py
@@ -1,1 +1,0 @@
-# utils package for cache_manager and future utilities
